# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EconomicsBody": {
        "properties": {
          "margin_pct": {
            "description": "Share of sales that is profit, in percent",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Margin Pct",
            "type": "number"
          },
          "truck_cost_per_mile": {
            "description": "What one truck trip costs per mile, in dollars",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Truck Cost Per Mile",
            "type": "number"
          }
        },
        "required": [
          "truck_cost_per_mile",
          "margin_pct"
        ],
        "title": "EconomicsBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Refusal": {
        "properties": {
          "line": {
            "title": "Line",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "line",
          "reason"
        ],
        "title": "Refusal",
        "type": "object"
      },
...
      "RowsBody": {
        "properties": {
          "mapping": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 5000,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "RowsBody",
        "type": "object"
      },
...
      "UploadBody": {
        "properties": {
          "csv": {
            "anyOf": [
              {
                "maxLength": 2000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The whole CSV file as text, with the column names on the first line",
            "title": "Csv"
          },
          "mapping": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Your column name for each StormSense column, if the names differ",
            "title": "Mapping"
          },
          "xlsx_base64": {
            "anyOf": [
              {
                "maxLength": 4000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "An Excel file (.xlsx), base64 encoded. Columns are matched by name",
            "title": "Xlsx Base64"
          }
        },
        "title": "UploadBody",
        "type": "object"
      },
      "UploadResult": {
        "properties": {
          "feed": {
            "title": "Feed",
            "type": "string"
          },
          "ignored_columns": {
            "items": {
              "type": "string"
            },
            "title": "Ignored Columns",
            "type": "array"
          },
          "kept": {
            "title": "Kept",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "missing_columns": {
            "items": {
              "type": "string"
            },
            "title": "Missing Columns",
            "type": "array"
          },
          "refusals": {
            "description": "The first 20 refused rows, with the line number and the reason",
            "items": {
              "$ref": "#/components/schemas/Refusal"
            },
            "title": "Refusals",
            "type": "array"
          },
          "refused": {
            "title": "Refused",
            "type": "integer"
          }
        },
        "required": [
          "feed",
          "kept",
          "refused",
          "refusals",
          "missing_columns",
          "ignored_columns",
          "message"
        ],
        "title": "UploadResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "StormSense",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ingest/economics": {
      "get": {
        "operationId": "get_economics_api_ingest_economics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomicsBody"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Economics",
        "tags": [
          "your data"
        ]
      },
      "put": {
        "description": "Truck cost per mile and product margin, used for the estimated profit of the plan.",
        "operationId": "set_economics_api_ingest_economics_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EconomicsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomicsBody"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Economics",
        "tags": [
          "your data"
        ]
      }
    },
...
    "/api/ingest/feeds/{feed}": {
      "delete": {
        "operationId": "clear_api_ingest_feeds__feed__delete",
        "parameters": [
          {
            "in": "path",
            "name": "feed",
            "required": true,
            "schema": {
              "title": "Feed",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Clear Api Ingest Feeds  Feed  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear",
        "tags": [
          "your data"
        ]
      }
    },
    "/api/ingest/feeds/{feed}/preview": {
      "get": {
        "operationId": "preview_api_ingest_feeds__feed__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "feed",
            "required": true,
            "schema": {
              "title": "Feed",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Preview Api Ingest Feeds  Feed  Preview Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "your data"
        ]
      }
    },
    "/api/ingest/feeds/{feed}/rows": {
      "post": {
        "description": "Send rows from another system as JSON. Same checks as a file upload.",
        "operationId": "rows_in_api_ingest_feeds__feed__rows_post",
        "parameters": [
          {
            "in": "path",
            "name": "feed",
            "required": true,
            "schema": {
              "title": "Feed",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rows In",
        "tags": [
          "your data"
        ]
      }
    },
    "/api/ingest/feeds/{feed}/template.xlsx": {
      "get": {
        "description": "An Excel template. Store and product codes are dropdowns, filled from the stores and products you uploaded.",
        "operationId": "template_xlsx_api_ingest_feeds__feed__template_xlsx_get",
        "parameters": [
          {
            "in": "path",
            "name": "feed",
            "required": true,
            "schema": {
              "title": "Feed",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Template Xlsx",
        "tags": [
          "your data"
        ]
      }
    },
    "/api/ingest/feeds/{feed}/upload": {
      "post": {
        "operationId": "upload_api_ingest_feeds__feed__upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "feed",
            "required": true,
            "schema": {
              "title": "Feed",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "your data"
        ]
      }
    },
...
  }
}